requirements.txt
visdom==0.2.4
seaborn==0.12.2
albumentations==0.4.6
imgaug==0.4.0
pycocotools==2.0.6
opencv-python==4.7.0.72
tqdm==4.65.0
torchnet==0.0.4
pandas==1.1.5
map-boxes==1.0.5
jupyter==1.0.0
